keep ExtraHashView in ckb-types
