+++ util/gen-types/src/core/mod.rs
@@ -1,8 +1,2 @@
 mod types;
 pub use types::*;
-
-#[cfg(feature = "calc-hash")]
-mod view;
-
-#[cfg(feature = "calc-hash")]
-pub use view::*;
