+++ util/types/src/core/mod.rs
@@ -34,7 +34,9 @@ pub use fee_rate::FeeRate;
 pub use reward::{BlockEconomicState, BlockIssuance, BlockReward, MinerReward};
 pub use transaction_meta::{TransactionMeta, TransactionMetaBuilder};
 pub use tx_pool::TransactionWithStatus;
-pub use views::{BlockView, HeaderView, TransactionView, UncleBlockVecView, UncleBlockView};
+pub use views::{
+    BlockView, ExtraHashView, HeaderView, TransactionView, UncleBlockVecView, UncleBlockView,
+};
 
 pub use ckb_gen_types::core::*;
 pub use ckb_occupied_capacity::{
