--- util/types/src/extension/mod.rs
@@ -1,10 +1,0 @@
-//! Extensions for packed bytes wrappers.
-//!
-//! Add methods for packed bytes wrappers.
-//!
-//! ### Warning
-//!
-//! No definitions for structs, enums is allowed.
-
-mod shortcuts;
-pub use shortcuts::*;
